fix(test): move BuildChainData test to L2G — requires game DLL for XRL.MarkovChain
The test calls BuildChainData/GenerateSentence which internally resolve
XRL.MarkovChain via reflection. This type only exists in Assembly-CSharp.dll,
so the test belongs in L2G behind #if HAS_GAME_DLL, not L1.
