+++ Mods/QudJP/Assemblies/QudJP.Tests/L1/MarkovCorpusTranslationPatchTests.cs
@@ -50,20 +50,4 @@ public sealed class MarkovCorpusTranslationPatchTests
         });
     }
 
-    [Test]
-    public void BuildChainData_ProducesJapaneseSentenceFromProductionCorpus()
-    {
-        var (order, corpusText) = MarkovCorpusTranslationPatch.LoadJapaneseCorpusSource();
-        var chainData = MarkovCorpusTranslationPatch.BuildChainData(corpusText, order);
-        var sentence = MarkovCorpusTranslationPatch.GenerateSentence(chainData);
-
-        Assert.Multiple(() =>
-        {
-            Assert.That(order, Is.EqualTo(2));
-            Assert.That(MarkovCorpusTranslationPatch.GetOpeningWordCount(chainData), Is.GreaterThan(50));
-            Assert.That(MarkovCorpusTranslationPatch.GetTransitionCount(chainData), Is.GreaterThan(500));
-            Assert.That(sentence, Is.Not.Empty);
-            Assert.That(MarkovCorpusTranslationPatch.ContainsJapaneseCharacters(sentence), Is.True);
-        });
-    }
 }
